==> Code/01_Data_Collection/02_State_Bills_Data.ipynb <==
import pandas as pd
import zipfile
import os
import shutil

# About original data

The Bills data is from the Open States [website](https://openstates.org/data/session-csv/). Open States is an organization that aggregates, standardizes, and cleans legislative data for all 50 states. The data used for this study is from the bulk data they offer of proposed bills in the state's legislature. The data is stored in zip files for each legislative session by state. In this codebook, we will aggregate all the states bills and sponsorship data from the zip files and save it as a csv file. Open States scrapes their data directly from governemtn websites and seems to be quite reliable. 

# Read in Data for Every State

Except for Nebraska as it has a unique state legislature.

#List of states, will help us when retrieving data from zipfiles
states = [ 'AK', 'AL', 'AR', 'AZ', 'CA', 'CO', 'CT', 'DC', 'DE', 'FL', 'GA',
           'HI', 'IA', 'ID', 'IL', 'IN', 'KS', 'KY', 'LA', 'MA', 'MD', 'ME',
           'MI', 'MN', 'MO', 'MS', 'MT', 'NC', 'ND', 'NH', 'NJ', 'NM',
           'NV', 'NY', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX',
           'UT', 'VA', 'VT', 'WA', 'WI', 'WV', 'WY']

The next code block does the following:

    1. Instatiates DataFrames that will hold our bills and sponsorships data.
    2. Loop through each state to extract wanted data:
       a. Unzip every .zip file for that state
       b. Read in wanted data into DataFrames from CSVs
       c. Delete the extracted data after reading everything in to not use too much memory.
       
    Result: Should end up with bills and sponsorships data for every state from the 2017-2018 sessions.

#1.
bills = pd.DataFrame() #instantiate dataframe to store the bills data for each state
sponsors = pd.DataFrame() #instantiate dataframe to store the sponsors data for each bill
abstracts = pd.DataFrame() #Instantiate abstracts dataframe

#2.
#Loop through each state as we have data for each state in seperate folders in the Data/Bills_Data folder
for state in states:
    print(f'Downloading {state} data') #Message showing which state we are downloading from
    
    state = state + '/' #End of path for location of the zip file (different for each state)
    path = '../../Data/Bills_Data/' + state #The whole path to the zipfiles for that particular state
    
    
    
    #2.a
    zip_end = '.zip' #We only want to extract files from zip files
    
    for file in os.listdir(path): #Go through every file in the state folder
        if file.endswith(zip_end): #If file is a zip file, do the next lines of code
            file_path = path+file #Path to zip file
            zip_object = zipfile.ZipFile(file_path) #Instantiate zip object for the particular file
            zip_object.extractall('../../Data/Bills_Data/') #Extract files from zip into the state folder
          
        
      
    #2.b
    #We now have extracted many files from the zip file
    #We only need information from the files ending in 'bills.csv' and 'bill_sponsorships.csv'.
    bills_end = 'bills.csv' #Ending used for bills info
    sponsors_end = 'bill_sponsorships.csv' #Ending used for sponsors data
    abstract_end = 'abstracts.csv' #Ending used for abstract data
    
    #Go through each root, directory, and file from this path(mainly just want files extracted from zip)
    for origin,sub,files in os.walk(path): 
        for file in files: #Go through each file found in the os.walk
            if file.endswith(bills_end): #If file is the 'blabla...bills.csv', read in that data
                new_data = pd.read_csv(origin+'/'+file) #Path to file
                bills = pd.concat([bills,new_data]) #Append new data read in to our bill dataframe instantiated above
            elif file.endswith(sponsors_end): #If file is the 'blabla...bill_sponsorships.csv', read in that data
                new_data = pd.read_csv(origin+'/'+file) #Path to file
                sponsors = pd.concat([sponsors,new_data]) #Append new data read in to our sponsor dataframe instantiated above
            elif file.endswith(abstract_end): #If file is the 'blabla...abstracts.csv', read in that data
                new_data = pd.read_csv(origin+'/'+file) #Path to file
                abstracts = pd.concat([abstracts,new_data]) #Append new data read in to our sponsor dataframe instantiated above
      
    #2.c
    #We have the data we want from this state. No need to keep unneeded data taking up memory.
    #Next few lines delete the extracted files from the zip
    #Go through each root, directory, and file from this path(only need the subdirectories)
    for origin,subs,files in os.walk(path): 
        for sub in subs: #Go through each newly created folder from unzipping state data
            if sub.endswith('.ipynb_checkpoints') == False: #Sometimes this shows up, chossing to ignore it
                shutil.rmtree(path + sub) #Remove that folder and everything in it, we still have the zip file though

# Data Inspection

bills.head()

sponsors.head()

abstracts.head()

bills.info()

sponsors.info()

abstracts.info()

There are some null values but overall, I am happy with this data extraction. Seems like we have 227,105 usable bills and many sponsor names that hopefully we will be able to use.

# Saving Data

bills.to_csv('../../Data/Bills_Data/state_bills_2017_2018.csv.zip', index = False)
sponsors.to_csv('../../Data/Bills_Data/bill_sponsors_2017_2018.csv.zip', index = False)
abstracts.to_csv('../../Data/Bills_Data/bill_abstracts_2017_2018.csv.zip', index = False)
