==> Code/01_Data_Collection/03_Bill_Progression_Data.ipynb <==
import pandas as pd
import zipfile
import os
import shutil

# About original data

The Legislative Bills Progression data comes from the Harvard Dataverse and was published by Dr. Garlick in April, 2023 ([link to data](https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/8PTHXT)). The dataset contains over a million bills for all states from the years of 2011 to 2019. It includes the 23 step progression of each bill showing where it failed or if it was enacted. Dr. Garlick and his team used Open Sates to create this data set which will work well for us because we will be able to connect it to the bill data we gathered. In this codebook, we will just filter the dataset to only include bills from the years of 2017 and 2018.

# Read in data

#Instantiate zip object
zip_object = zipfile.ZipFile('../../Data/Legislative_Progression/dataverse_files.zip') 
#Extract all files in zip file
zip_object.extractall('../../Data/Legislative_Progression/unzipped')

#Save the codebook
codebook = pd.read_csv('../../Data/Legislative_Progression/unzipped/ajo_lpcode_codebook_v0.1.csv')
#Save the actual data
bill_progression = pd.read_csv('../../Data/Legislative_Progression/unzipped/ajo_lpcode_v0.1.csv')

#Delete our extracted data so we don't waste memory
shutil.rmtree('../../Data/Legislative_Progression/unzipped')

The warning let's us know that some columns have mixed data types. We will deal with this problem in the data cleaning phase.

# Filter data and save

#We only need bills made after 2017
bill_progression_filtered = bill_progression[bill_progression['year1'] >= 2017]

#Save data to compressed file
bill_progression_filtered.to_csv('../../Data/Legislative_Progression/bill_progression_2017_2018.csv.zip', index = False)

codebook.to_csv('../../Data/Legislative_Progression/codebook.csv.zip', index = False)
