==> Code/01_Data_Collection/05_Combining_Data.ipynb <==
import pandas as pd

# Merging datasets

We know have different datasets in seperate files. We want to merge them all together into one dataframe to be used for our modeling. During this process, we are bound to lose some data that does not match up with the other datasets. As long as we do not lose too many data points, that will be okay.

# Read in Data

bills = pd.read_csv('../../Data/Bills_Data/state_bills_2017_2018.csv.zip')
abstracts = pd.read_csv('../../Data/Bills_Data/bill_abstracts_2017_2018.csv.zip')
progression = pd.read_csv('../../Data/Legislative_Progression/bill_progression_2017_2018.csv.zip')
codebook = pd.read_csv('../../Data/Legislative_Progression/codebook.csv.zip')
leg_control_2017 = pd.read_csv('../../Data/Legislative_Control_Data/legis_control_2017.csv')
leg_control_2018 = pd.read_csv('../../Data/Legislative_Control_Data/legis_control_2018.csv')
bills_sponsor_party = pd.read_csv('../../Data/Bills_Data/bills_sponsor_party.csv.zip')

# Merging bills with bill's progression data

We will need to make a common id that we can merge the datsets on. As both datasets were derived from the common source of Open States, this should not be too difficult. The progression data has a bill_id colum and so does the bills data. The values in this colum look like: "HB 1", "HB 102", "SB 1001" etc. The letters represent where the bill was made and the numbers represent when the bill was made. So, the first bill introduced in the Colorado House for a particular session would be "HB 1", but this is also the case for the first bill introduced in any state as well. All in all, we need more than just the bill identifier to merge the two datasets. 

We will also need to use the session identifier for both datasets. This looks different sate by state but identifies which legislative session the bill was introduced in. Just in case that two states use the same name for a legislative session, for example: "2017 session", we will also make sure the states match up too.

Thus, to merge the two datasets, we will make a custom id for each row that will consist of the State abbreviation, the legisaltive session and the bill id. We will then be able to match both datsets together.

#Dictionary that converts sate names to abbreviations , needed for bills data
us_state_to_abbrev = {"Alabama": "AL", "Alaska": "AK", "Arizona": "AZ", "Arkansas": "AR", "California": "CA",
    "Colorado": "CO", "Connecticut": "CT", "Delaware": "DE", "Florida": "FL", "Georgia": "GA", "Hawaii": "HI",
    "Idaho": "ID", "Illinois": "IL", "Indiana": "IN", "Iowa": "IA", "Kansas": "KS", "Kentucky": "KY", "Louisiana": "LA",
    "Maine": "ME", "Maryland": "MD", "Massachusetts": "MA", "Michigan": "MI", "Minnesota": "MN", "Mississippi": "MS",
    "Missouri": "MO", "Montana": "MT", "Nebraska": "NE", "Nevada": "NV", "New Hampshire": "NH", "New Jersey": "NJ", "New Mexico": "NM",
    "New York": "NY","North Carolina": "NC","North Dakota": "ND","Ohio": "OH","Oklahoma": "OK", "Oregon": "OR",
    "Pennsylvania": "PA", "Rhode Island": "RI", "South Carolina": "SC", "South Dakota": "SD", "Tennessee": "TN", "Texas": "TX", "Utah": "UT", 
    "Vermont": "VT", "Virginia": "VA", "Washington": "WA", "West Virginia": "WV", "Wisconsin": "WI", "Wyoming": "WY", 
    "District of Columbia": "DC", "American Samoa": "AS", "Guam": "GU", "Northern Mariana Islands": "MP", "Puerto Rico": "PR", 
    "United States Minor Outlying Islands": "UM", "U.S. Virgin Islands": "VI"
}

#Making custom ID for consisting of state, session name and bill id for the progression dataset
progression['custom_id'] = progression['state'] + progression['sess_str'] + progression['bill_id']

#Formatting bills data to match progression data
bills['jurisdiction'] = bills['jurisdiction'].map(us_state_to_abbrev) #Convert state names into abbreviations
bills['jurisdiction'] = bills['jurisdiction'].map(lambda x: str(x).lower()) #undercase the abbreviations

#Making the custom ID for bills data
bills['custom_id'] = bills['jurisdiction'] +  bills['session_identifier'] + bills['identifier']

df = bills.merge(progression, how = 'inner', on = 'custom_id') #Merge on the custom id and leave out rows that don't match

df.head(1)

print(f'Shape of Bills data: {bills.shape} \nShape of Bill\'s Progression data: {progression.shape}\nShape of merged data: {df.shape}')

We lossed many rows, but we still have almost 160,000. This will be enough for our study.

# Merging with the legislative control data

For this merge, we will make our own custom id again. It will just consist of the year and state of each row.

#Dictionary that converts sate names to abbreviations , needed to use plotly graph
us_state_to_abbrev = {"Alabama": "AL", "Alaska": "AK", "Arizona": "AZ", "Arkansas": "AR", "California": "CA",
    "Colorado": "CO", "Connecticut": "CT", "Delaware": "DE", "Florida": "FL", "Georgia": "GA", "Hawaii": "HI",
    "Idaho": "ID", "Illinois": "IL", "Indiana": "IN", "Iowa": "IA", "Kansas": "KS", "Kentucky": "KY", "Louisiana": "LA",
    "Maine": "ME", "Maryland": "MD", "Massachusetts": "MA", "Michigan": "MI", "Minnesota": "MN", "Mississippi ": "MS",
    "Missouri": "MO", "Montana": "MT", "Nebraska": "NE", "Nevada": "NV", "New Hampshire": "NH", "New Jersey": "NJ", "New Mexico": "NM",
    "New York": "NY","North Carolina": "NC","North Dakota  ": "ND","Ohio": "OH","Oklahoma": "OK", "Oregon": "OR",
    "Pennsylvania": "PA", "Rhode Island ": "RI", "South Carolina": "SC", "South Dakota": "SD", "Tennessee": "TN", "Texas": "TX", "Utah": "UT", 
    "Vermont": "VT", "Virginia": "VA", "Washington": "WA", "West Virginia": "WV", "Wisconsin": "WI", "Wyoming": "WY", 
    "District of Columbia": "DC", "American Samoa": "AS", "Guam": "GU", "Northern Mariana Islands": "MP", "Puerto Rico": "PR", 
    "United States Minor Outlying Islands": "UM", "U.S. Virgin Islands": "VI"
}

#Formatting the control data to match progression data for both years
leg_control_2017['states'] = leg_control_2017['states'].map(us_state_to_abbrev) #Convert state names into abbreviations
leg_control_2017['states'] = leg_control_2017['states'].map(lambda x: str(x).lower()) #undercase the abbreviations
leg_control_2017['year'] = 2017 #add year colum

leg_control_2018['states'] = leg_control_2018['states'].map(us_state_to_abbrev) #Convert state names into abbreviations
leg_control_2018['states'] = leg_control_2018['states'].map(lambda x: str(x).lower()) #undercase the abbreviations
leg_control_2018['year'] = 2018 #add year colum

leg_control = pd.concat([leg_control_2017, leg_control_2018]) #Combine both datasets into one

#Make custom id to match up with legislative control data
df['custom_id2'] = df['state'] + df['year1'].map(lambda x: str(x))

#Do the same for control data
leg_control['custom_id2'] = leg_control['states'] + leg_control['year'].map(lambda x: str(x))

df = df.merge(leg_control, how = 'inner', on = 'custom_id2')

df.head(1)

print(f'Shape of merged data: {df.shape}')

# Merging with abstracts data

Although most bills don't have data with thier abstracts, we will merge in what we have.

added = df.merge(abstracts, how = 'inner', left_on= 'id', right_on= 'bill_id').shape[0]
print(f'we are only addding {added} abstracts to our dataset.')

df = df.merge(abstracts, how = 'left', left_on= 'id', right_on= 'bill_id') 

# Merging Legislator Sponsor data

This is a straight forward merge, we just merge on the bill ids.

df_with_sponsor = df.merge(bills_sponsor_party, how = 'inner', left_on = 'id_x', right_on = 'bill_id')

print(f'Much data was lossed. With the sponsorship data we only have {df_with_sponsor.shape[0]} rows.')
print(f'Without the sponsorship data we have {df.shape[0]} rows.')

I will save both dataframes in seperate files. 

# Saving Data

df_with_sponsor.to_csv('../../Data/Merged_Data/merged_data_with_sponsorship.csv.zip')
df.to_csv('../../Data/Merged_Data/merged_data.csv.zip')

df
